string literals, function definitions, and packages with ::
... sort of. ANTLR prioritizes lexer rules in order of definition. As
such, you generally want to go from specific to general in the lexer.
However, if a more specific rule matches, the more general one will not.

This means that a parser rule which only references the general one will
not match, I think. As such, this commit adds some alternatives to
parser rules so that a more specific rule matching is still OK.