bunch of mystical lexer changes
Done with intent: the Q ops are _not_ in the group with higher priority
in the Guacamole BNF. I had included them with all the other ops in the
top of the ANTLR Lexer. This gave them much more priority than they were
meant to have.

Other changes are more or less guesswork and trying to reproduce the
Guacamole BNF (mostly by reordering rules to respect priority).