Cope with lexer ambiguity for varIdent
I think ANTLR is a little less capable of dealing with ambiguity than
Marpa. As a result, the Perl BNF has a number of rules which partially
overlap, like SubName, VarIdent, and SubNameNonQLike.

In ANTLR, as far as I can tell, this means that whichever rule is first
in the lexer grammar and matches the given sequence will match. This
means that varIdentExpr will fail to match if the variable name is also
a valid SubName.

Judging from StackOverflow (https://stackoverflow.com/a/57163873) it's
pretty normal to expand the set of lexer rules that a parse rule accepts
in order to tolerate such ambiguity. So that's what this commit does! :)

(Note for future readers: my current level of ANTLR expertise is "hit it
with a stick until my test file parses", so please take any assertions
like "little less capable" with an entire salt mine)