Note about overriding builtins
Keywords and SubNames collide. In order for keywords to ever lex, they
need to come first in the grammar. However, in order for people to be
able to write 'sub if { }' the parser needs to accept 'if' in
a subNameExpr.

In order to do this I think the way to go is to add an alternative to
subNameExpr which allows whatever set of keywords may be sub names (all
of them?)